Update patch set 5
Patch Set 5:

(12 comments)

I looked through most of this CL.  Haven't finished loopschedchecks.go yet.

I think this currently touches too much of the compiler to be safe for 1.8.  It would be ok normally, but not at this point in the cycle.  I want it to be obvious that none of these changes affect anything when the experiment is off.
Have loopschedchecks do all the stuff you're currently caching from deadstore, then you don't need to modify deadstore at all.

Patch-set: 5
Label: Code-Review=0