Update patch set 2
Patch Set 2:

> TryBots are happy.

I feel like this isn't the right fix for this bug.

The code generated after walk processes the CALLPART looks ok to me:

.   CONVNOP l(18) tc(1) FUNC-func()
.   .   PTRLIT l(18) esc(no) tc(1) PTR64-*struct { F uintptr; R X }
.   .   .   STRUCTLIT l(18) esc(no) tc(1) STRUCT-struct { F uintptr; R X }
.   .   .   .   TYPE <S> l(18) tc(1) implicit(true) type=PTR64-*struct { F uintptr; R X } PTR64-*struct { F uintptr; R X }
.   .   .   STRUCTLIT-list
.   .   .   .   STRUCTKEY l(18) x(0) main.F
.   .   .   .   .   CFUNC l(18) tc(1) uintptr
.   .   .   .   .   .   NAME-main.(X).Print-fm u(1) a(true) l(18) x(0) class(PFUNC) tc(1) used(true) FUNC-func()

.   .   .   .   STRUCTKEY l(18) x(8) main.R
.   .   .   .   .   STRUCTLIT l(18) tc(1) main.X
.   .   .   .   .   .   TYPE main.X u(1) a(true) l(5) x(0) class(PEXTERN) tc(1) type=main.X main.X main.X
.   .   .   .   .   STRUCTLIT-list
.   .   .   .   .   .   STRUCTKEY l(18) x(0) main.A
.   .   .   .   .   .   .   SLICELIT l(18) esc(h) tc(1) SLICE-[]byte
.   .   .   .   .   .   .   .   LITERAL-0 u(1) a(true) l(18) untyped number
.   .   .   NAME-main..autotmp_5 u(1) a(true) l(18) x(0) class(PAUTO) esc(N) tc(1) used(true) STRUCT-struct { F uintptr; R X }

The STRUCTLIT for X has a STRUCTKEY for A but not for B, but that is fine, just as if I had written that in source code.  Somehow B is not getting initialized by subsequent rewrites.  I think the temp variable fix in this CL is just papering over a larger issue in STRUCTLIT processing (potentially triggered in more ways than the CALLPART path demonstrated in #18410).

Patch-set: 2