Update patch set 2
Patch Set 2:

> (1 comment)

I've simplified the logic. The Tx.done comment was correct and shouldn't be modified, I just moved setting the Tx.done to the beginning of Commit and rollback to reflect what was already documented. I can't add test right now due to time constraints, but I should have time on Thursday to do so to attempt a test.

Patch-set: 2
Label: Code-Review=0