cmd/compile: check pointer validity on conversion
The runtime will check validity of pointers internally and panic on
occasion, so ensure that these pointers cannot be introduced in the
first place. This will only happen on conversion to unsafe.Pointer from
non-pointer types, so most internal runtime pointer mangling should be
free from these checks.

This change requires that existing "abuses" of unsafe.Pointer to store
non-pointer values are removed. Therefore, the runtime mmap interface
(which encodes an error as a value < 4096) has been changed to uintptr,
and the conversion to unsafe.Pointer only occurs after the value has
been verified to be valid. An alternate version of this change would
encode two return values for the low-level mmap functions, but this
seems sufficient to avoid ambiguity. As a result, a related TODO was
removed from the cgo_mmap hook.

For performance reasons, these checks are normally disabled when
building the runtime package. The expectation is that unsafe pointer
manipulation should be rare outside this package which defines core
types. The checks are enabled for the runtime package when building in
MSAN mode or the race detector is enabled.

Change-Id: I778a4e2d4b6089308abd4b7d7b87acb3836d5fa1