Update patch set 2
Patch Set 2:

Thanks for the explanation. Assigning to temp indeed triggers zeroing. Is there any other place that we generate struct literal without assignment? Maybe closure?

This seems to generate more temps than needed. For example, if the parent node is already OAS, this also generates a temp (along with another zeroing and a copy). (but how to detect it?)

Patch-set: 2