Update patch set 1
Patch Set 1:

Why check this with explicit code?

I don't see why this is helpful except to help debugging
bad pointer problems to catch the exact location that
introduced the bad pointer.

The other problem is that this check is a strict subset
of check enforced by the GC. For example, this won't
catch pointers to unused objects/spans.

Also, this check is not strong enough as there are
checks for p < _PageSize in the runtime and _PageSize
is currently 8KB.

Anyway, I think we need more rationale behind this change.
Is there any discussion that I have missed?

Patch-set: 1