Update patch set 1
Patch Set 1:

> Feel free to push back on this until the backstory & use case is explained to your satisfaction in the commit message and/or a linked bug.

Yeah, I'd like to see an issue filed with more complete information.

If it's just for debugging bad pointer issues, then we can even
make a runtime call to check if the pointer really is a good pointer.

Another potential solution is to have the write barrier always enabled
when the check is active. Then we don't need any compiler instrumentation
at all -- just do full pointer validity check in the writer barrier.
With the added benefit that we don't even need to recompile a program
to enable the extensive pointer check, which is probably beneficial
for production use of this feature.

Patch-set: 1