Update patch set 1
Patch Set 1:

> Why check this with explicit code?
 > 
 > I don't see why this is helpful except to help debugging
 > bad pointer problems to catch the exact location that
 > introduced the bad pointer.

Right -- the idea is to prevent the bad pointers from being introduced in the first place, not explode randomly at a later point in time when the stack is copied and/or GC occurs (which can be very difficult to debug).

Stopping a program as soon as it becomes invalid strikes me as important, not just useful for debugging. If you could detect arbitrary memory corruption or runtime errors, you would certainly want to stop the program immediately, and this error is being thrown as the result of a later internal runtime check.

 > The other problem is that this check is a strict subset
 > of check enforced by the GC. For example, this won't
 > catch pointers to unused objects/spans.
 
Can you point me to those checks? If there's an easy class of invalid pointer that would cause runtime errors and could also be checked then I will try to add these.

 > Also, this check is not strong enough as there are
 > checks for p < _PageSize in the runtime and _PageSize
 > is currently 8KB.

Huh, this is really odd. 8kb isn't the page size, so it seems quite surprising to have whatever it is be named _PageSize. (And it prompts the question -- are those checks attempting to check against the actual page size, or whatever _PageSize is?)

IMO, the checks against _PageSize are not valid with _PageSize != minZeroPage. Checks against minZeroPage would be perfectly valid, since those pointers are explicitly disallowed and the implied contract seems to be only that minZeroPage is unmapped. A pointer to 0x1002 could be perfectly valid.

That all said, it would also be reasonable to change this to refer to _PageSize if there's a lot of existing runtime checks and that's the consensus. The important thing is that I want to stop executing as soon as the runtime will consider the program invalid.

 > Anyway, I think we need more rationale behind this change.
 > Is there any discussion that I have missed?

Yes.

I have a program that deals with unsafe pointers (for a good reason, and safely -- let's put that aside for the purposes of this discussion).

There was effectively a case with:
addr := 1
x := (int*)(unsafe.Pointer(uintptr(addr)))

And the code ran exactly as expected, except it exploded in copystack due to the pointer sanity check. So, to address this issue, I posted:
https://go-review.googlesource.com/#/c/34663/

After discussion there and offline, I now understand why these pointers are checked and considered invalid. But I still don't like the idea of things exploding randomly since it's nearly impossible to debug.

So instead, this is an attempt to fix the other side of the problem: if that code is indeed invalid and _should_ cause the runtime to panic, then that should happen there, not at some random later point.

Patch-set: 1