Update patch set 1
Patch Set 1:

> > Feel free to push back on this until the backstory & use case is
 > explained to your satisfaction in the commit message and/or a
 > linked bug.
 > 
 > Yeah, I'd like to see an issue filed with more complete
 > information.
 > 
 > If it's just for debugging bad pointer issues, then we can even
 > make a runtime call to check if the pointer really is a good
 > pointer.

This is something I could use personally, but I'm already covered (as the check is trivial). I'm trying to prevent this same problem from affecting others and cover cases that I potentially haven't thought about.

 > Another potential solution is to have the write barrier always
 > enabled
 > when the check is active. Then we don't need any compiler
 > instrumentation
 > at all -- just do full pointer validity check in the writer
 > barrier.
 > With the added benefit that we don't even need to recompile a
 > program
 > to enable the extensive pointer check, which is probably beneficial
 > for production use of this feature.

This sounds good. I have no idea what the best way to implement such a feature is, I just stumbled around until I had something that did what I wanted :)

The check in writebarrierptr does indeed seem to be what we want. (And actually, it does seem to imply that the other runtime checks should be going against minPhysPageSize _not_ _PageSize, as noted above.)

Due to checks in copystack and GC, the mere existence of such a pointer on the stack or elsewhere needs to be prevented. That's why I added the check on conversion from uintptr to unsafe.Pointer. Is there a good way to hook in to the writebarrier function in only this case? (Pardon my ignorance)

Patch-set: 1