Update patch set 3
Patch Set 3:

Please revert. This is wrong. The output from nm is not valid input to -X. The output from nm has mangled import paths in the symbol names, and the input to -X does not. (See CL 31970 and issue #16710.)

The doc string already says '-X importpath.name'. If clarification is needed, one of those needs to be explained better, although honestly I don't see which.

Patch-set: 3