Update patch set 13
Patch Set 13:

I fixed the buildrun test mode to use a temporary directory, but I also (based on last-year discussions with Austin) lowered the iteration count to 1021, so that a 97 nS loop would check every 100 microseconds for a reschedule request.  Ideally we would be cleverer than that, but decrementing from 32k meant that we'd generally wait well longer to check for a rescheduling request.  

And because I changed that, I am re-running benchmarks on AMD64 and PPC64 to be sure that timings don't change too much.

Patch-set: 13