cmd/compile: optimize non-empty-interface type conversions
When doing i.(T) for non-empty-interface i and concrete type T,
there's no need to read the type out of the itab. Just compare the
itab to the itab we expect for that interface/type pair.

Also optimize type switches by putting the type hash of the
concrete type in the itab. That way we don't need to load the
type pointer out of the itab.

DO NOT REVIEW
DO NOT SUBMIT

Update #18492
Fixes #18595

Change-Id: I49e280a21e5687e771db5b8a56b685291ac168ce