Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > How about this fix instead?
> > 
> > --- a/src/cmd/compile/internal/gc/sinit.go
> > +++ b/src/cmd/compile/internal/gc/sinit.go
> > @@ -1078,6 +1078,7 @@ func anylit(n *Node, var_ *Node, init *Nodes) {
> >  
> >                 var r *Node
> >                 if n.Right != nil {
> > +                       init.Append(nod(OAS, n.Right, nil))
> >                         r = nod(OADDR, n.Right, nil)
> >                         r = typecheck(r, Erv)
> >                 } else {
> > 
> > This seems more likely to me to fix the bug no matter how the &struct{} is generated.
> 
> That appears to work, but I don't understand why.  RHS of PTRLIT is appropriate for LHS of assignment?

The n.Right of PTRLIT, if not nil, contains a temporary that is used as the backing store for the PTRLIT.  It was not zeroed (unlike the other branch of this "if" which allocates a new, and hence zeroed, value).
It would equally be ok to put the zeroing at the point where n.Right is set (escape analysis?).

Patch-set: 2