Update patch set 3
Patch Set 3:

As a sanity check, I can run this CL on our large regression test to see if it breaks anything. That should improve confidence that it's the right fix for Go 1.8.

Let me know when the CL is ready and I'll kick off the test. It takes about a day to run.

Patch-set: 3