runtime: add definitions for SetGoroutineLabels and Do
This change defines runtime/pprof.SetGoroutineLabels and runtime/pprof.Do, which
are used to set profiler labels on goroutines. The change defines functions
in the runtime for setting and getting profile labels, and sets and unsets
profile labels when goroutines are created and deleted. The change also adds
the package runtime/internal/proflabel, which defines the structure the runtime
uses to store profile labels.

NOTE TO REVIEWERS: This change's tests don't pass. I'm working on debugging them.

Change-Id: I747a4400141f89b6e8160dab6aa94ca9f0d4c94d