Update patch set 1
Patch Set 1:

> This doesn't even compile.

I will confess to not trying to compile it, but I copied an existing file and changed the number. Currently waiting on an m68k gccgo bootstrap build to test.

 > m68k isn't a valid GOOS, and you don't have a +build tag either.

+build isn't needed when the architecture is in the file name, at least on gccgo. See getrandom_linux_amd64/386/alpha.go for examples.

 > What is the motivation for this?

Supporting getrandom in gccgo on m68k. It could live in gccgo only, but there are other architectures that aren't supported by the real native go which have definitions here.

Patch-set: 1