Update patch set 1
Patch Set 1:

One of the (implicit) rationale for reserving a GOOS or GOARCH
here is that there is non-negligible chance that gc will work
on such platforms.

Therefore, we left out alpha and ia64 from the list either
because there're no longer commercially available or deemed
too hard to write a reasonably good compiler for.

It's ok for gccgo to support more platforms than what is available
in go/build/syslist.go. One reason for reserving GOOS and GOARCH
here is to make sure gc and gccgo use the identical name for
the same platform to make writing portable code easier.

That is, only add m68k if you plan to add m68k support to gc.

Patch-set: 1