Update patch set 1
Patch Set 1:

> Therefore, we left out alpha and ia64 from the list either
 > because there're no longer commercially available or deemed
 > too hard to write a reasonably good compiler for.

Alpha is still fully supported in Debian.

 > It's ok for gccgo to support more platforms than what is available
 > in go/build/syslist.go. One reason for reserving GOOS and GOARCH
 > here is to make sure gc and gccgo use the identical name for
 > the same platform to make writing portable code easier.
 > 
 > That is, only add m68k if you plan to add m68k support to gc.

That's not unlikely to happen in the future. I just recently made Firefox work on m68k :). People are working on FPGA implementations of these older CPUs and m68k is, in fact, quite advanced being one of the few architectures which support Compare-and-Swap with two independent arguments, for example.

Patch-set: 1