Update patch set 1
Patch Set 1:

This is not the place to decide it, but I do feel that it is appropriate to reserve GOARCH values for architectures that will only be implemented by compilers other than gc.  We don't want to go overboard but we do want to add popular architectures.

Currently the names in the gccgo version of go/build/syslist.go that are not in the gc version are alpha and m68k.

Patch-set: 1
Label: Code-Review=0