Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is not the place to decide it, but I do feel that it is appropriate to reserve GOARCH values for architectures that will only be implemented by compilers other than gc.  We don't want to go overboard but we do want to add popular architectures.

The problem is, gccgo is not the only "other" compiler.
If we relax the policy, how can we reject other additions
for architectures only implemented by some other compilers?

I'm fine with extending the policy, as long as we set a
clear criteria for new reserved GOOSes and GOARCHes that
are only supported by toolchains other than gc.

I think the problem is worse with GOOS as there are far
larger number of OS than the number of architectures
that can reasonably run Go.

Patch-set: 1