[dev.garbage] runtime: Speedup systemstack
Currently systemstack check whether it is running
on the signal stack then whether it is already running
on the systemstack before finally dropping into the common
case code of switching from the g (normal) stack
to the g0 (system) stack. Adjust the code to check
if you are on the g stack and if so immediately jump
to the code that moves you to the g0 stack. This is
particularly important for ROC since it makes many
more systemstack calls than the non-ROC code.

Change-Id: I8be2fb53f0ba98a070a7995039d513f82f2973a7