[dev.typealias] spec: use 'embedded type' rather than 'anonymous' or 'embedded' field
First steps towards defining type aliases in the spec.
This is a nomenclature clarification, not a language change.

The spec used all three terms 'embedded type', 'anonymous field',
and 'embedded field'. Users where using the terms inconsistently.

The notion of an 'anonymous' field was always misleading since they
always had a de-facto name. With type aliases that name becomes even
more important because we may have different names for the same type.

Struct fields are always 'embedded' in a struct; thus speaking of an
'embedded field' doesn't seem distinguishing enough.

'Embedded type' expresses the quality of such fields best: They
are still fields, with a name (given by the type) and they consist
of a type directly embedded in a struct.

For #18130.

Change-Id: I2083bbc85788cab0b2e2cb1ff58b2f979491f001