[dev.garbage] runtime: Avoid sweep of free spans
When ROC recycles a span it may discover that the
span is empty. In this case ROC will return
the span to the allocator for general reuse.
Unfortunately, the span is in heap_.sweepSpans
array of spans to be swept after the next GC
cycle. This confuses the sweeper when it
encounters an invalid span state.

This CL is a place holder that fixes the problem
by having the sweeper correctly process spans that
are not in the appropriate state. This will
eventually be changed by ROC removing the span
from mheap_.sweepSpans.

Change-Id: I85fce80e3c2ce1d2ee4e65027efc948ee89eaf46