Update patch set 1
Patch Set 1:

> I think we need to consider the reflect package.  Currently
 > reflect.StructField has a bool field Anonymous, with the doc
 > comment "is an embedded field".  There is a panic in
 > reflect.StructOf: "reflect.StructOf: illegal anonymous field type
 > ".  There are various comments that say "anonymous", though those
 > could be changed.  There are also comments referring to "embedded
 > fields" and "embedded types".
 > 
 > Anyhow, the one thing we can't change is reflect.StructField.Anonymous.

Agreed - that's just unfortunate. I think in the spec, for the purpose of clarity, we should nevertheless settle on embedded type or embedded field, simply because these fields actually have (always had) a name and that name is becoming more important.

I suggest we leave this one as is for the moment and I'll produce the remaining spec CL(s) for type aliases and then we can make a final decision at the end.

There may be other stuff coming up that may sway us one way or the other.

Patch-set: 1