Update patch set 1
Patch Set 1:

(2 comments)

Thanks for making this change.

This is certainly an improvement over "anonymous", but "embedded type" suggests that it's a property of the type, not the field.  The special quality of embedded fields is that their subfields and methods are "promoted", or made available with shortcut names, to the enclosing struct type.  

I was surprised to discover (during the 1.8 alias work) that there is no necessary technical reason why only anonymous fields should be subject to promotion.  If you had some syntax to indicate it, you could in principle allow multiple fields to be promoted, independent of their names and types.

If we were choosing terms anew, "promoted field" would capture this quality best---the term "promoted" already appears in the spec---but unfortunately this term is not widely used in the community.

BTW: typo in CL description: s/Users where/Users were/.

Patch-set: 1
Label: Code-Review=0