Update patch set 1
Patch Set 1:

> PS: I'm wondering if it might be better to have the debug check explicit?

You mean a top level flag, not hidden behind -d? I'm not sure why or who would use it. We should probably have a builder that runs with it enabled, though, just as we have an SSA check builder.


> And leave testdclstack() in place at least after each file as a basic security measure for now?

Sure. So that I don't screw it up, which call to testdclstack is that?

Patch-set: 1