Update patch set 1
Patch Set 1:

> > PS: I'm wondering if it might be better to have the debug check
 > explicit?
 > 
 > You mean a top level flag, not hidden behind -d? I'm not sure why
 > or who would use it. We should probably have a builder that runs
 > with it enabled, though, just as we have an SSA check builder.
 > 

Sorry for being unclear. No, I meant doing the "debug_dclstack == 0" test explicitly at call sites where we invoke testdclstack. We already check (in several places) for syntax errors before we call it.

It's a bigger change but it may be surprising to call testdclstack and not seing anything happen. It's not pretty, but perhaps better to be explicit at each call site than have another surprise.

 > 
 > > And leave testdclstack() in place at least after each file as a
 > basic security measure for now?
 > 
 > Sure. So that I don't screw it up, which call to testdclstack is
 > that?

It's the call in noder.go: parseFile (line 37). Perhaps it makes sense to do this one unconditionally because the stack should be empty here unless I'm missing something.

In any case would be good to verify that this doesn't screw up the compile time again.

What do you think?

Patch-set: 1