Update patch set 1
Patch Set 1:

This error is something very similar to some of the debugging I faced with plugins on macs. I believe what's happening here is because we have two versions of the function iface_a.init in the process, the dynamic relocation for the function pointer in the symtab of the plugin is being pointed to the function from the original program, which means the symtab contains a function pointer into an unexpected range.

Somewhere, I had a fix for this for macs that I didn't push, because all it did was expose another bug I couldn't whack. I'm digging through my local git repos to see if something will jog my memory.

Patch-set: 1