Update patch set 1
Patch Set 1:

> > Patch Set 1:
 > >
 > > > > Patch Set 1:
 > >  > > Somewhere, I had a fix for this for macs that I didn't push,
 > >  > because all it did was expose another bug I couldn't whack.
 > I'm
 > >  > digging through my local git repos to see if something will
 > jog my
 > >  > memory.
 > >  >
 > >  > I don't really understand all the implications of the "Check a
 > PC
 > >  > near but not at the very end." comment in moduledataverify1.
 > Maybe
 > >  > the stuff that code is doing is not valid for plugins with a
 > very
 > >  > small number of functions?
 > >
 > > I thought you wrote that code. :-)
 > 
 > Heh no I just copied it around. Somehow I suspect Russ wrote it.
 > 
 > > I can safely say I don't understand why that "near by not at the
 > very end" is there. If you don't either, it's probably wrong. Maybe
 > the only problem here is that verification code.
 > 
 > It would be interesting to just comment that block out and see if
 > anything else falls over I guess.

Keith's itabs test in this CL passes if the moduledataverify1 call is removed.

I was working on the assumption that the verification logic is valid, and it was uncovering a problem that would bite us later. But if we are both at sea in moduledataverify1, it's just as likely a bug in the verification code.

I'm not really sure what the right thing to do for 1.8 is. Maybe we should remove the moduledataverify1 call? Or some part of it?

Patch-set: 1