Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > > Patch Set 1:
>  > >
>  > > > > Patch Set 1:
>  > >  > > Somewhere, I had a fix for this for macs that I didn't push,
>  > >  > because all it did was expose another bug I couldn't whack.
>  > I'm
>  > >  > digging through my local git repos to see if something will
>  > jog my
>  > >  > memory.
>  > >  >
>  > >  > I don't really understand all the implications of the "Check a
>  > PC
>  > >  > near but not at the very end." comment in moduledataverify1.
>  > Maybe
>  > >  > the stuff that code is doing is not valid for plugins with a
>  > very
>  > >  > small number of functions?
>  > >
>  > > I thought you wrote that code. :-)
>  > 
>  > Heh no I just copied it around. Somehow I suspect Russ wrote it.
>  > 
>  > > I can safely say I don't understand why that "near by not at the
>  > very end" is there. If you don't either, it's probably wrong. Maybe
>  > the only problem here is that verification code.
>  > 
>  > It would be interesting to just comment that block out and see if
>  > anything else falls over I guess.
> 
> Keith's itabs test in this CL passes if the moduledataverify1 call is removed.
> 
> I was working on the assumption that the verification logic is valid, and it was uncovering a problem that would bite us later. But if we are both at sea in moduledataverify1, it's just as likely a bug in the verification code.
> 
> I'm not really sure what the right thing to do for 1.8 is. Maybe we should remove the moduledataverify1 call? Or some part of it?

Well I drank a coffee and got distracted by this and think I know what's happening. The test in moduledataverify1 is testing, for the last 4 functions in the module, that the pc->data tables cover from the start of the function until a bit before the start of the next function in the functab. Buuuut in plugins, functions matching "onlycsymbol" are not written into the functab. So if you check a function that happens to be laid out just before an "onlycsymbol" function, the code tries to check that the pc->data covers all the way until to the function after the "onlycsymbol" function, which it obviously doesn't. I suspect if you set debugPcln to true in the runtime, all plugins would fail to load in this way, Keith's is just being unlucky for some reason.

So this is just a validation problem. We could disable this particular check for plugins, perhaps? Not sure how to do that. I also don't really understand why these functions are excluded from the ftab.

Patch-set: 1