[dev.typealias] spec: define regular and alias type declarations
Now that a type declaration can declare both a new named type and
alternatively, a type alias for a given type, the notion of
a named type is becoming problematic. Given a name for a type, it
may or may not be a "named type" in the original sense; yet clearly
the type has a name and thus can be considered "named".

In this first attempt of the spec change for type aliases, I have
tried to remedy the issue by introducing new terminology. I am not
wedded to the particular terms used, but I do feel somewhat strongly
that simply re-using the existing term "named type" is going to be
very confusing, particularly to people new to Go who are unfamiliar
with the history.

I am aware that this will cause some pain for existing users. However,
if we assume that Go adoption is continuing to grow, it is safe to say
that there will be many more new users to Go who won't notice the
nomenclature change, vs. an increasingly shrinking percentage of people
familiar with the old terminology. It seems better to bite the bullet
now. The important part is that there is no language change besides
the ability to give types names without introducing a new type.

Specific changes:

- Introduced the notion of a "branded type": This is what formerly
  was called a "named type". As said above, a better term is welcome
  but I think we need a specific term. It should be a simple find/replace
  if we decide on a better term.

- Adjusted language that implicitly assumed a named type is different
  from an unnamed type: A type with a name may now be identical to
  a type literal if the type name is a type alias denoting that literal.
  Furthermore, two differently named types may denote the same branded
  type.

- Adjusted type declarations: There are now two forms, one of which
  creates a branded type first before the type name is bound to it.

Type declarations are now more in line with all the other declarations:
It's primary purpose is to bind a name to a type. In the specific (old)
form, it also creates a new branded (formerly) named type before binding
the name.

There is a bit of a mindset change here, which is necessary now that
we have type aliases: Methods are attached to the branded type, not
the type name. This reflects reality (in the implementation) and is
the primary reason why I believe we should introduce a new term to
refer to that type.

Change-Id: I54e3c488c90e6880c8bb55e9a13d8dc985914d4e